double reporting and index local progres.

+++ arangod/ClusterEngine/ClusterCollection.cpp
@@ -201,7 +201,7 @@ uint64_t ClusterCollection::numberDocuments(transaction::Methods* trx) const {
 
 std::shared_ptr<Index> ClusterCollection::createIndex(
     velocypack::Slice info, bool restore, bool& created,
-    std::shared_ptr<std::function<arangodb::Result(uint64_t)>> progress) {
+    std::shared_ptr<std::function<arangodb::Result(double)>> progress) {
   TRI_ASSERT(ServerState::instance()->isCoordinator());
 
   // prevent concurrent dropping
