+++ arangod/VocBase/LogicalCollection.cpp
@@ -1081,7 +1081,7 @@ std::shared_ptr<Index> LogicalCollection::lookupIndex(VPackSlice info) const {
 
 std::shared_ptr<Index> LogicalCollection::createIndex(
     VPackSlice info, bool& created,
-    std::shared_ptr<std::function<arangodb::Result(uint64_t)>> progress) {
+    std::shared_ptr<std::function<arangodb::Result(double)>> progress) {
   auto idx = _physical->createIndex(info, /*restore*/ false, created,
                                     std::move(progress));
   if (idx) {
